+++ lib/yandex_money/api.rb
@@ -1,4 +1,5 @@
 require "yandex_money/api/version"
+require "yandex_money/config"
 require "yandex_money/exceptions"
 require "yandex_money/logger/default"
 require "yandex_money/logger/empty"